Move configurations to src/config
