+++ client/src/config/configGenerator.js
