+++ client/src/config/registerServiceWorker.js
